Create DependenciesHelpers file, add tests
