+++ Library/Homebrew/cmd/uses.rb
@@ -7,6 +7,7 @@
 require "formula"
 require "cli/parser"
 require "cask/caskroom"
+require "dependencies_helpers"
 
 module Homebrew
   extend DependenciesHelpers
